moved battery stuff into the battery module
